CKEditor + delete confirmations

+++ frontend/src/blog/PostDetail.jsx
@@ -1,14 +1,16 @@
 import { Link, useLoaderData, redirect, useNavigate } from "react-router-dom";
 import { deletePost } from "../http";
+import parse from "html-react-parser";
 
 export default function PostDetail() {
   const navigate = useNavigate();
   const post = useLoaderData();
 
   async function handleDeletePost(post_id) {
-    const resData = await deletePost(post_id);
-
-    navigate("/blog");
+    if (confirm("Are you sure to delete post?")) {
+      const resData = await deletePost(post_id);
+      navigate("/blog");
+    }
   }
 
   return (
@@ -16,8 +18,10 @@ export default function PostDetail() {
       <div className="row">
         <div className="col">
           <h3>{post.title}</h3>
-          <img src={post.imageUrl} alt="react-img" className="img-fluid" />
-          <p>{post.body}</p>
+          <p>
+            <img src={post.imageUrl} alt="react-img" className="img-fluid" />
+          </p>
+          {parse(post.body)}
           <p>
             {post.author} | {post.published}
           </p>
@@ -26,7 +30,7 @@ export default function PostDetail() {
               Back
             </Link>{" "}
             <Link
-              className="btn btn-sm btn-primary"
+              className="btn btn-sm btn-secondary"
               to={`/blog/${post.id}/edit`}
             >
               Edit
